# backend/swagger.yaml
openapi: 3.0.0
info:
  title: Student Record Management System API
  version: 1.0.0
servers:
  - url: http://localhost:3000/api
tags:
  - name: Auth
  - name: Admin  
  - name: Student
paths:
  /auth/student/register:
    post:
      tags: [Auth]
      summary: Student registration
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name: { type: string }
                email: { type: string }
                password: { type: string }
      responses:
        '201':
          description: Success
  /auth/login:
    post:
      tags: [Auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email: { type: string }
                password: { type: string }
      responses:
        '200':
          description: Success
  /admin/students:
    post:
      tags: [Admin]
      security: [{ bearerAuth: [] }]
      summary: Create student
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, email, age, gender, department, semester, enrollmentNumber, address]
              properties:
                name:
                  type: string
                  example: "John Doe"
                email:
                  type: string
                  format: email
                  example: "john.doe@example.com"
                age:
                  type: integer
                  example: 20
                gender:
                  type: string
                  enum: [Male, Female, Other]
                  example: "Male"
                department:
                  type: string
                  example: "Computer Science"
                semester:
                  type: integer
                  example: 6
                enrollmentNumber:
                  type: string
                  example: "CS2021001"
                address:
                  type: string
                  example: "123 Main St, City, State 12345"
      responses:
        '201':
          description: Success
    get:
      tags: [Admin]
      security: [{ bearerAuth: [] }]
      summary: Get all students
      responses:
        '200':
          description: Success
  /admin/students/{enrollmentNumber}:
    get:
      tags: [Admin]
      security: [{ bearerAuth: [] }]
      summary: Get student by enrollment number
      parameters:
        - name: enrollmentNumber
          in: path
          required: true
          schema: { type: string }
          example: "CS2021001"
      responses:
        '200':
          description: Success
    put:
      tags: [Admin]
      security: [{ bearerAuth: [] }]
      summary: Update student by enrollment number
      parameters:
        - name: enrollmentNumber
          in: path
          required: true
          schema: { type: string }
          example: "CS2021001"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name: { type: string, example: "John Doe" }
                email: { type: string, example: "john.doe@example.com" }
                age: { type: integer, example: 20 }
                gender: { type: string, enum: [Male, Female, Other], example: "Male" }
                department: { type: string, example: "Computer Science" }
                semester: { type: integer, example: 6 }
                address: { type: string, example: "123 Main St, City, State" }
      responses:
        '200':
          description: Success
    delete:
      tags: [Admin]
      security: [{ bearerAuth: [] }]
      summary: Delete student by enrollment number
      parameters:
        - name: enrollmentNumber
          in: path
          required: true
          schema: { type: string }
          example: "CS2021001"
      responses:
        '200':
          description: Success
  /admin/marks:
    post:
      tags: [Admin]
      security: [{ bearerAuth: [] }]
      summary: Create marks for student
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [enrollmentNumber, subjects]
              properties:
                enrollmentNumber:
                  type: string
                  example: "CS2021001"
                subjects:
                  type: array
                  items:
                    type: object
                    properties:
                      name: { type: string, example: "Mathematics" }
                      internal: { type: number, example: 18 }
                      external: { type: number, example: 72 }
                      total: { type: number, example: 90 }
      responses:
        '201':
          description: Success
  /admin/marks/{enrollmentNumber}:
    get:
      tags: [Admin]
      security: [{ bearerAuth: [] }]
      summary: Get student marks
      parameters:
        - name: enrollmentNumber
          in: path
          required: true
          schema: { type: string }
          example: "CS2021001"
      responses:
        '200':
          description: Success
    put:
      tags: [Admin]
      security: [{ bearerAuth: [] }]
      summary: Update student marks
      parameters:
        - name: enrollmentNumber
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                subjects:
                  type: array
                  items:
                    type: object
                    properties:
                      name: { type: string, example: "Mathematics" }
                      internal: { type: number, example: 18 }
                      external: { type: number, example: 72 }
                      total: { type: number, example: 90 }
      responses:
        '200':
          description: Success
  /student/profile:
    get:
      tags: [Student]
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: Success
  /student/scorecard:
    get:
      tags: [Student]
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: Success
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT